requirements.txt
# This file is used by pip to install required python packages
# Command: pip install -r requirements.txt

Flask==1.0.2
mongoengine==0.18.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
flask-bcrypt==0.7.1
flask-login==0.4.1
Flask-WTF==0.14.2
passlib==1.7.2
# To connect to online MongoDB atlas we need dnspython module....
dnspython==1.16.0
